move admin webpacker-controlled js to admin/ subdir
