+++ app/javascript/packs/application.js
@@ -7,6 +7,6 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-import '../src/js/member_sortable';
-import '../src/js/tab_selection_in_anchor';
-import '../src/js/simple_uppy_file_input';
+import '../src/js/admin/member_sortable';
+import '../src/js/admin/tab_selection_in_anchor';
+import '../src/js/admin/simple_uppy_file_input';
