+++ app/javascript/src/js/admin/simple_uppy_file_input.js
