+++ app/javascript/src/js/admin/tab_selection_in_anchor.js
